+++ src/BaGet.Core.Server/Extensions/IServiceCollectionExtensions.cs
@@ -1,0 +1,31 @@
+using BaGet.Configuration;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Cors.Infrastructure;
+using Microsoft.AspNetCore.Http.Features;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Options;
+
+namespace BaGet.Core.Server.Extensions
+{
+    public static class IServiceCollectionExtensions
+    {
+        public static IServiceCollection ConfigureHttpServices(this IServiceCollection services)
+        {
+            services
+                .AddMvc()
+                .AddApplicationPart(typeof(BaGet.Controllers.PackageController).Assembly)
+                .SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+
+            services.AddCors();
+            services.AddSingleton<IConfigureOptions<CorsOptions>, ConfigureCorsOptions>();
+            services.AddSingleton<IConfigureOptions<ForwardedHeadersOptions>, ConfigureForwardedHeadersOptions>();
+            services.Configure<FormOptions>(options =>
+            {
+                options.MultipartBodyLengthLimit = int.MaxValue;
+            });
+
+            return services;
+        }
+    }
+}
